+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-### v0.1.1
+### v0.2.0
 
 - add changelog
 - add link to putrecord-test repository
-- re-introduce proper White Wind handling
+- re-introduce proper WhiteWind handling
+- add WhiteWind tests
 
 ### v0.1.0